requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
distro==1.9.0
Django==5.2.3
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jiter==0.10.0
openai==1.86.0
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stripe==12.2.0
text-unidecode==1.3
tomli==2.2.1
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
zope.interface==7.2
